Add some basic pot handling
Ideally we'd be looking the pot name up in the API but this code makes
no direct calls to the API -- we're manually exporting. Will probably
loop back on this at some future point and tighten it up, but likely not
until the API has matured more. Would be nice to e.g. auto open
beancount accounts as sub Monzo accounts.